+++ bridgekeeper/rules.py
@@ -416,14 +416,27 @@ class In(Rule):
 #: This rule is satisfied by any group the user is in.
 in_current_groups = In(lambda user: user.groups.all())
 
+# from https://groups.google.com/forum/#!msg/django-developers/jEkCdzGnzRE/6lJQV4AqCwAJ
+
+
+def add_prefix(q_obj, prefix):
+    return Q(
+        *(
+            add_prefix(child, prefix)
+            if isinstance(child, Q)
+            else (prefix + '__' + child[0], child[1])
+            for child in q_obj.children
+        ),
+        _connector=q_obj.connector,
+        _negated=q_obj.negated,
+    )
+
 
 class Relation(Rule):
     """Check that a rule applies to a ForeignKey.
 
     :param attr: Name of a foreign key attribute to check.
     :type attr: str
-    :param model: Model class on the other side of the foreign key.
-    :type model: type
     :param rule: Rule to check the foreign key against.
     :type rule: Rule
 
@@ -432,25 +445,22 @@ class Relation(Rule):
     access the related applicant::
 
         perms['foo.view_application'] = Relation(
-            'applicant', Applicant, perms['foo.view_applicant'])
+            'applicant', perms['foo.view_applicant'])
     """
 
-    def __init__(self, attr, model, rule):
+    def __init__(self, attr, rule):
         self.attr = attr
-        self.model = model
         self.rule = rule
 
     def __repr__(self):
-        return "Relation({!r}, {}, {!r})".format(
-            self.attr, self.model.__name__, self.rule)
+        return "Relation({!r}, {!r})".format(
+            self.attr, self.rule)
 
     def query(self, user):
-        # Unfortunately you can't use Q objects on a relation, only proper
-        # QuerySets.
         related_q = self.rule.query(user)
         if related_q is UNIVERSAL or related_q is EMPTY:
             return related_q
-        return Q(**{self.attr + '__in': self.model.objects.filter(related_q)})
+        return add_prefix(related_q, self.attr)
 
     def check(self, user, instance=None):
         if instance is None:
@@ -482,8 +492,6 @@ class ManyRelation(Rule):
         set :attr:`~django.db.models.ForeignKey.related_name` or
         :attr:`~django.db.models.ForeignKey.related_query_name`.
     :type query_attr: str
-    :param model: Model class on the other side of the relationship.
-    :type model: type
     :param rule: Rule to check the foreign object against.
     :type rule: Rule
 
@@ -492,19 +500,18 @@ class ManyRelation(Rule):
     relationship with their agency::
 
         perms['foo.view_customer'] = ManyRelation(
-            'agencies', Agency, Is(lambda user: user.agency))
+            'agency_set', 'agency', Is(lambda user: user.agency))
     """
 
-    def __init__(self, attr, query_attr, model, rule):
+    def __init__(self, attr, query_attr, rule):
         # TODO: add support for 'all' as well as 'exists'
         self.attr = attr
         self.query_attr = query_attr
-        self.model = model
         self.rule = rule
 
     def __repr__(self):
-        return "ManyRelation({!r}, {}, {}, {!r})".format(
-            self.attr, self.query_attr, self.model.__name__, self.rule)
+        return "ManyRelation({!r}, {!r}, {!r})".format(
+            self.attr, self.query_attr, self.rule)
 
     def query(self, user):
         # Unfortunately you can't use Q objects on a relation, only proper
@@ -512,7 +519,7 @@ class ManyRelation(Rule):
         related_q = self.rule.query(user)
         if related_q is UNIVERSAL or related_q is EMPTY:
             return related_q
-        return Q(**{self.query_attr + '__in': self.model.objects.filter(related_q)})
+        return add_prefix(related_q, self.query_attr)
 
     def check(self, user, instance=None):
         if instance is None: